+++ Tests/Topaz.Tests.AzureCLI/AppServiceKuduTests.cs
@@ -10,7 +10,7 @@ public class AppServiceKuduTests : TopazFixture
         var subscriptionId = string.Empty;
         await RunAzureCliCommand("az group create -n rg-kudu-deploy -l westeurope");
         await RunAzureCliCommand("az appservice plan create -n plan-kudu-deploy -g rg-kudu-deploy --sku B1 -l westeurope");
-        await RunAzureCliCommand("az webapp create -n kudu-cli-deploy -g rg-kudu-deploy --plan plan-kudu-deploy");
+        await RunAzureCliCommand("az webapp create -n kudu-cli-deploy -g rg-kudu-deploy --plan plan-kudu-deploy --runtime \"PYTHON:3.14\"");
         await RunAzureCliCommand("az account show", response =>
         {
             subscriptionId = response["id"]!.GetValue<string>();
@@ -42,7 +42,7 @@ public class AppServiceKuduTests : TopazFixture
     {
         await RunAzureCliCommand("az group create -n rg-kudu-list -l westeurope");
         await RunAzureCliCommand("az appservice plan create -n plan-kudu-list -g rg-kudu-list --sku B1 -l westeurope");
-        await RunAzureCliCommand("az webapp create -n kudu-cli-list -g rg-kudu-list --plan plan-kudu-list");
+        await RunAzureCliCommand("az webapp create -n kudu-cli-list -g rg-kudu-list --plan plan-kudu-list --runtime \"PYTHON:3.14\"");
 
         var subscriptionId = string.Empty;
         await RunAzureCliCommand("az account show", response =>
@@ -71,9 +71,12 @@ public class AppServiceKuduTests : TopazFixture
             {
                 var array = response.AsArray();
                 Assert.That(array, Is.Not.Null);
-                Assert.That(array!.Count, Is.GreaterThanOrEqualTo(1));
-                Assert.That(array[0]!["status"]!.GetValue<string>(), Is.EqualTo("succeeded"));
-                Assert.That(array[0]!["deployer"]!.GetValue<string>(), Is.EqualTo("Push Deployer"));
+                Assert.Multiple(() =>
+                {
+                    Assert.That(array, Is.Not.Empty);
+                    Assert.That(array[0]!["status"]!.GetValue<string>(), Is.EqualTo("succeeded"));
+                    Assert.That(array[0]!["deployer"]!.GetValue<string>(), Is.EqualTo("Push Deployer"));
+                });
             });
 
         await RunAzureCliCommand("az group delete -n rg-kudu-list --yes");